working rotating boundary nodes. Walls still not moving

+++ cylinder_2D.hpp
@@ -6,7 +6,7 @@ class cylinder_2D : public geometry_2D {
 
 public:
 
-    cylinder_2D(lb::coordinate<int> centerOfMass, double orientation, int radius) :
+    cylinder_2D(lb::coordinate<double> centerOfMass, double orientation, int radius) :
             geometry_2D(centerOfMass,orientation), mRadius(radius) {
         update_boundary_and_internal_nodes();
     }
